TeraskyRag/requirements.txt
# RAG and API dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
chromadb>=0.4.0
py2neo>=2021.2.4
python-dotenv>=1.0.0
boto3>=1.26.0
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0

# Additional RAG-specific dependencies
streamlit>=1.28.0
requests>=2.31.0